Move query_sql to reader, then refactor benches to superfile and supertable only.
