+++ benches/utils/mod.rs
@@ -6,7 +6,7 @@
 // - `corpus/` — synthetic data (stream + optional grading cache)
 // - `ingest/` — stream corpus → append → commit → object storage
 // - `fixture/` — one shared 10M ingest per process (`supertable_all`)
-// - `fts_superfile`, `vector_superfile` — 1M superfile bench bodies
+// - `superfile`, `supertable` — tier-specific bench runners by modality
 // - `tiers`, `markdown`, `rss` — storage backends + reporting
 
 pub mod corpus;
@@ -18,9 +18,8 @@ pub mod report;
 pub mod rss;
 pub mod tiers;
 
-pub mod fts_superfile;
-pub mod sql_bench;
+pub mod superfile;
+pub mod supertable;
+
 pub mod sql_diag;
-pub mod supertable_bench;
 pub mod unified_object_store;
-pub mod vector_superfile;
