+++ src/supertable/writer.rs
@@ -396,7 +396,7 @@ impl SupertableWriter {
         // writer session are not yet in the manifest.
         let supertable = Supertable::from_inner(Arc::clone(&self.inner));
         let target_ids = supertable
-            .scan_ids_matching(predicate)
+            .reader().scan_ids_matching(predicate)
             .map_err(MutationError::PredicateEval)?;
         let matched = target_ids.len();
         if matched > MAX_TARGETS_PER_MUTATION {
@@ -471,7 +471,7 @@ impl SupertableWriter {
         // writer's buffer don't count toward the match set.
         let supertable = Supertable::from_inner(Arc::clone(&self.inner));
         let target_ids = supertable
-            .scan_ids_matching(predicate)
+            .reader().scan_ids_matching(predicate)
             .map_err(MutationError::PredicateEval)?;
         let matched = target_ids.len();
         if matched > MAX_TARGETS_PER_MUTATION {
