+++ tests/supertable/storage/smoke_s3.rs
@@ -353,7 +353,7 @@ async fn supertable_smoke_via_s3_wire_protocol() {
     let pre = cache.stats();
     assert_eq!(pre.n_cold_fetches, 0);
     let batches = consumer
-        .query_sql("SELECT COUNT(*) AS n FROM supertable")
+        .reader().query_sql("SELECT COUNT(*) AS n FROM supertable")
         .expect("query_sql via S3");
     assert_eq!(batches.len(), 1);
     let post = cache.stats();
@@ -622,7 +622,7 @@ async fn supertable_tvfs_through_query_sql_via_s3_wire_protocol() {
     //    in exactly two titles ("alpha vector one", "alpha
     //    vector two"), so the TVF must return >= 2 rows.
     let bm25 = consumer
-        .query_sql(&format!(
+        .reader().query_sql(&format!(
             "SELECT _id FROM bm25_search('title', 'alpha', {BM25_TOP_K})"
         ))
         .expect("bm25_search via query_sql over S3");
@@ -635,7 +635,7 @@ async fn supertable_tvfs_through_query_sql_via_s3_wire_protocol() {
     // 2. vector_search through query_sql. k=3.
     let vec_sql = format!("SELECT _id FROM vector_search('emb', '{q_csv}', 3)");
     let vector = consumer
-        .query_sql(&vec_sql)
+        .reader().query_sql(&vec_sql)
         .expect("vector_search via query_sql over S3");
     assert!(
         count_rows(&vector) >= 1,
@@ -647,7 +647,7 @@ async fn supertable_tvfs_through_query_sql_via_s3_wire_protocol() {
     let hybrid_sql =
         format!("SELECT _id FROM hybrid_search('title', 'alpha', 'emb', '{q_csv}', 5)");
     let hybrid = consumer
-        .query_sql(&hybrid_sql)
+        .reader().query_sql(&hybrid_sql)
         .expect("hybrid_search via query_sql over S3");
     let hyb_rows = count_rows(&hybrid);
     assert!(
